# contract/openapi.yml
openapi: 3.0.0
info:
  title: IT Corpo API
  version: 2.0.0
  description: >
    API for managing IT corporation resources.

    For full documentation, please visit [the readme](https://github.com/ducin-public/itcorpo-api).
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /employees:
    get:
      summary: List all employees
      operationId: getEmployees
      tags:
        - Employees
        - Pagination
        - Search
      parameters:
        - $ref: '#/components/parameters/EmployeeName'
        - $ref: '#/components/parameters/EmployeeDepartmentId'
        - $ref: '#/components/parameters/EmployeeSkills'
        - $ref: '#/components/parameters/EmployeeSkillsFiltering'
        - $ref: '#/components/parameters/EmployeeSalaryFrom'
        - $ref: '#/components/parameters/EmployeeSalaryTo'
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '400':
          description: >
            Invalid employees search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    post:
      summary: Create a new employee
      operationId: createEmployee
      tags:
        - Employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: >
            Invalid employee input request body
            @see {@link EmployeeInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Employee with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid data state (e.g. department doesn't exist, office not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /employees/count:
    get:
      summary: Get total number of employees
      operationId: getEmployeesCount
      tags:
        - Employees
        - Search
      parameters:
        - $ref: '#/components/parameters/EmployeeName'
        - $ref: '#/components/parameters/EmployeeDepartmentId'
        - $ref: '#/components/parameters/EmployeeSkills'
        - $ref: '#/components/parameters/EmployeeSkillsFiltering'
        - $ref: '#/components/parameters/EmployeeSalaryFrom'
        - $ref: '#/components/parameters/EmployeeSalaryTo'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1250
        '400':
          description: >
            Invalid employees search criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /employees/{employeeId}:
    get:
      summary: Get employee by ID
      operationId: getEmployeeById
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          example: 91720
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    
    put:
      summary: Update employee
      operationId: updateEmployee
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: >
            Invalid employee input request body
            @see {@link EmployeeInput}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already taken by another employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid data state (e.g. department doesn't exist, office not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      summary: Delete employee
      operationId: deleteEmployee
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Employee deleted successfully
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete employee that is assigned to active projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /employees/{employeeId}/projects:
    get:
      summary: Get projects which employee is assigned to
      operationId: getEmployeeProjects
      description: >
        Returns a list of projects that the employee is assigned to.
      tags:
        - Employees
      parameters:
        - name: employeeId
          in: path
          required: true
          example: 91720
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectEmployeeInvolvement'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  responses:
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: The service is temporarily unavailable. Maybe there is maintenance?
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Money:
      type: number
      description: Monetary value in EUR
    DateString:
      type: string
      format: date-time
      description: ISO 8601 date-time string
    Email:
      type: string
      format: email
      description: Email address string
    Phone:
      type: string
      description: Phone number string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: An application-level error code (**not** HTTP status code)
        message:
          type: string
          description: Text description of the error that has occurred
      required:
        - message
...
    Nationality:
      type: string
      description: >
        Nationality of employee as an
        ISO 3166-1 alpha-2 country code
      enum:
        - "US"
        - "UK"
        - "FR"
        - "DE"
        - "NL"
        - "PL"
        - "IT"
        - "ES"
      example: "US"
    ContractType:
      type: string
      description: Type of employment contract
      enum:
        - "CONTRACT"
        - "PERMANENT"
      example: "PERMANENT"
    Skill:
      type: string
      description: Employee skill name
      example: "JavaScript"
    Employee:
      type: object
      required:
        - id
        - nationality
        - department
        - keycardId
        - account
        - salary
        - office
        - firstName
        - lastName
        - title
        - contractType
        - email
        - hiredAt
        - expiresAt
        - personalInfo
        - skills
        - bio
      properties:
        id:
          type: integer
          example: 91720
        nationality:
          $ref: '#/components/schemas/Nationality'
        department:
          type: string
        keycardId:
          type: string
        account:
          type: string
        salary:
          $ref: '#/components/schemas/Money'
        office:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: string
        firstName:
          type: string
        lastName:
          type: string
        title:
          type: string
        contractType:
          $ref: '#/components/schemas/ContractType'
        email:
          $ref: '#/components/schemas/Email'
        hiredAt:
          $ref: '#/components/schemas/DateString'
        expiresAt:
          $ref: '#/components/schemas/DateString'
        personalInfo:
          type: object
          required:
            - age
            - phone
            - email
            - dateOfBirth
            - address
          properties:
            age:
              type: integer
              minimum: 0
            phone:
              $ref: '#/components/schemas/Phone'
            email:
              $ref: '#/components/schemas/Email'
            dateOfBirth:
              $ref: '#/components/schemas/DateString'
            address:
              type: object
              required:
                - street
                - city
                - country
              properties:
                street:
                  type: string
                city:
                  type: string
                country:
                  type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        bio:
          type: string
        imgURL:
          type: string
      example:
        id: 1234
        nationality: "DE"
        department: Marketing
        keycardId: "KC-9876"
        account: "DE89 3704 0044 0532 0130 00"
        salary: 75000
        office: ["Berlin", "HQ"]
        firstName: "Hans"
        lastName: "Schmidt"
        title: "Senior Developer"
        contractType: "PERMANENT"
        email: "hans.schmidt@itcorpo.com"
        hiredAt: "2020-01-15T00:00:00.000Z"
        expiresAt: "2025-01-14T23:59:59.999Z"
        personalInfo:
          age: 35
          phone: "+49 123 456789"
          email: "hans.schmidt@gmail.com"
          dateOfBirth: "1988-05-20T00:00:00.000Z"
          address:
            street: "Alexanderplatz 1"
            city: "Berlin"
            country: "Germany"
        skills: ["JavaScript", "TypeScript", "React"]
        bio: "Experienced developer with focus on frontend technologies"
        imgURL: "hans-schmidt-profile.jpg"
    EmployeeInput:
      type: object
      required:
        - nationality
        - department
        - keycardId
        - account
        - salary
        - office
        - firstName
        - lastName
        - title
        - contractType
        - email
        - hiredAt
        - expiresAt
        - personalInfo
        - skills
        - bio
      properties:
        nationality:
          $ref: '#/components/schemas/Nationality'
        department:
          type: string
        keycardId:
          type: string
        account:
          type: string
        salary:
          $ref: '#/components/schemas/Money'
        office:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: string
        firstName:
          type: string
        lastName:
          type: string
        title:
          type: string
        contractType:
          $ref: '#/components/schemas/ContractType'
        email:
          $ref: '#/components/schemas/Email'
        hiredAt:
          $ref: '#/components/schemas/DateString'
        expiresAt:
          $ref: '#/components/schemas/DateString'
        personalInfo:
          type: object
          required:
            - age
            - phone
            - email
            - dateOfBirth
            - address
          properties:
            age:
              type: integer
              minimum: 0
            phone:
              $ref: '#/components/schemas/Phone'
            email:
              $ref: '#/components/schemas/Email'
            dateOfBirth:
              $ref: '#/components/schemas/DateString'
            address:
              type: object
              required:
                - street
                - city
                - country
              properties:
                street:
                  type: string
                city:
                  type: string
                country:
                  type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        bio:
          type: string
        imgURL:
          type: string
...
    ProjectStatus:
      type: string
      description: Status of the ongoing project's workflow
      enum:
        - "PLANNING"
        - "ACTIVE"
        - "COMPLETED"
        - "ON_HOLD"
      example: "ACTIVE"
    EngagementLevel:
      type: string
      description: Level of employee engagement in the project
      enum:
        - "FULL_TIME"
        - "PARTIAL_PLUS"
        - "HALF_TIME"
        - "ON_DEMAND"
      example: "FULL_TIME"
...
    ProjectEmployeeInvolvement:
      type: object
      description: Employee's involvement in a project
      required:
        - employeeId
        - projectId
        - employeeName
        - projectName
        - projectStatus
        - engagementLevel
        - since
      properties:
        employeeId:
          type: integer
        projectId:
          type: string
        employeeName:
          type: string
        projectName:
          type: string
        projectStatus:
          $ref: '#/components/schemas/ProjectStatus'
        engagementLevel:
          $ref: '#/components/schemas/EngagementLevel'
        since:
          type: string
          format: date
  parameters:
    PageNumber:
      name: _page
      in: query
      required: false
      description: Paginate data collections
      schema:
        type: number
        description: Page number to retrieve
        example: 1

    # Benefit parameters
...
    EmployeeName:
      name: employeeName
      in: query
      required: false
      description: Filter employees by name
      schema:
        type: string
        description: Filter employees by name
        example: "John Doe"
    EmployeeDepartmentId:
      name: departmentId
      in: query
      required: false
      description: Filter employees by department ID
      schema:
        type: string
        description: Filter employees by department ID
        example: "123"
    EmployeeSkills:
      name: skills
      in: query
      required: false
      description: Filter employees by skills
      schema:
        type: string
        description: Filter employees by skills according to `skillsFiltering`
        example: "JavaScript,React"
    EmployeeSkillsFiltering:
      name: skillsFiltering
      in: query
      required: false
      description: How to match employee skills
      schema:
        type: string
        enum: ["ANY", "ALL"]
        default: "ANY"
        description: If more than one skill is passed, return either employees with any of the skills (`ANY`) or with all of them (`ALL`)
        example: "ALL"
    EmployeeSalaryFrom:
      name: salaryFrom
      in: query
      required: false
      description: Minimum salary amount
      schema:
        type: string
        description: Minimum salary amount
        example: "5000"
    EmployeeSalaryTo:
      name: salaryTo
      in: query
      required: false
      description: Maximum salary amount
      schema:
        type: string
        description: Maximum salary amount
        example: "10000"

    # Office parameters
...